Continued work on unit tests ... Added - help, vcs, mail show - mail reply management (show, add, remove, change, default)
Redone the std* capture. Must be done via interception and redirection
of `::puts` itself.  To much inside of external packages (for.ex:
`cmdr::*`) we do not capture with just `m::msg`.

Considering to undo and remove `m::msg`. Defered at the moment, might
still be usable for the planned vcs notification callback thing in the
back of my mind.

Extended the low-level db layer with ability to reset database command
(close database and prepare for re-open).

Added db clearing to test support also.