Added helper application `gh-pull`.
Given a github user name, and (optionally) real name it pulls the list
of repositories for that user, and writes it as mirror import spec to
stdout, with description and language information.

Forks are resolved to their root repository.

After human review and filtering (removal of unwanted repositories,
changing names, descriptions, etc) this can then be directly imported
into a mirror instance.