Submission management work
- New app: `mirror-submit`. CGI helper.

- Database changes
  Table `submission`
  New columns
  - vcode	Capture VCS information
  - description	Capture description = name of mirror set to be
  - session	Generated

Extended `mirror submit` to capture vcs and description,
optional. When not specified it now auto-detects and generates this
information like `mirror add` does.

Extended `mirror accept` to enable override of vcs and
description. When not specified falls back to the information coming
from `mirror submit`, or fully generated like `mirror add`.

Refactored the mail generation in `mirror accept/reject` a bit, and
moved the common header/footer text into separate generator
procedures.

Moved the generation of repository description / mirror set name into
the VCS plugins. Extended the same in the github plugin to query
github for user name and project description.

Extended the submission logic to take the session id into account.

    Resubmission of an url in the same session overwrites the existing
    submission. This enables editing vcs, description, etc of a
    submission without creating many near-duplicates of it.

    The session of submissions made before the database change is
    `:lock:`. As this is not created by neither cli nor cgi these
    entries cannot be edited.

    The CLI generates session ids of the form `cli.<day>/host/user`.
    This enables a user on a specific host to edit all submissions
    from the current day. The next day they are locked and new entries
    are made under the new session.

    The CGI generates a session id of the form `(clicks)`, whenever
    the form is freshly entered. While reposting the form the session
    id is kept.

Added a new svg to the site generator, for the submit action of the
CGI helper.